Introduct AbstractSparseArray, start rewriting BlockSparseArray
