+++ NDTensors/src/lib/SparseArrayDOKs/src/SparseArrayDOKs.jl
@@ -1,11 +1,12 @@
 module SparseArrayDOKs
 include("defaults.jl")
-include("sparsearraydok.jl")
+include("abstractsparsearray.jl")
 include("sparsearrayinterface.jl")
 include("base.jl")
 include("broadcast.jl")
 include("map.jl")
 include("baseinterface.jl")
-include("conversion.jl")
+include("convert.jl")
 include("SparseArrayDOKsSparseArraysExt.jl")
+include("sparsearraydok.jl")
 end
