+++ NDTensors/src/lib/SparseArrayDOKs/src/SparseArrayDOKsSparseArraysExt.jl
@@ -2,4 +2,4 @@ using SparseArrays: SparseArrays
 using ..SparseArrayInterface: nstored
 
 # Julia Base `AbstractSparseArray` interface
-SparseArrays.nnz(a::SparseArrayDOK) = nstored(a)
+SparseArrays.nnz(a::AbstractSparseArray) = nstored(a)
