+++ NDTensors/src/lib/SparseArrayDOKs/src/abstractsparsearray.jl
@@ -1,0 +1,1 @@
+abstract type AbstractSparseArray{T,N} <: AbstractArray{T,N} end
