+++ NDTensors/src/lib/SparseArrayDOKs/src/baseinterface.jl
@@ -1,15 +1,15 @@
 using ..SparseArrayInterface: SparseArrayInterface
 
-Base.size(a::SparseArrayDOK) = a.dims
+Base.size(a::AbstractSparseArray) = error("Not implemented")
 
-function Base.getindex(a::SparseArrayDOK, I...)
+function Base.similar(a::AbstractSparseArray, elt::Type, dims::Tuple{Vararg{Int}})
+  return error("Not implemented")
+end
+
+function Base.getindex(a::AbstractSparseArray, I...)
   return SparseArrayInterface.sparse_getindex(a, I...)
 end
 
-function Base.setindex!(a::SparseArrayDOK, I...)
+function Base.setindex!(a::AbstractSparseArray, I...)
   return SparseArrayInterface.sparse_setindex!(a, I...)
-end
-
-function Base.similar(a::SparseArrayDOK, elt::Type, dims::Tuple{Vararg{Int}})
-  return SparseArrayDOK{elt}(undef, dims)
 end
