+++ NDTensors/src/lib/SparseArrayDOKs/src/convert.jl
@@ -1,0 +1,1 @@
+Base.convert(type::Type{<:AbstractSparseArray}, a::AbstractArray) = type(a)
