+++ NDTensors/src/lib/SparseArrayDOKs/src/sparsearraydok.jl
@@ -1,7 +1,7 @@
 using Dictionaries: Dictionary
 
 # TODO: Parametrize by `data`?
-struct SparseArrayDOK{T,N,Zero} <: AbstractArray{T,N}
+struct SparseArrayDOK{T,N,Zero} <: AbstractSparseArray{T,N}
   data::Dictionary{CartesianIndex{N},T}
   dims::NTuple{N,Int}
   zero::Zero
@@ -58,4 +58,36 @@ end
 
 function SparseArrayDOK{T}(::UndefInitializer, dims::Tuple{Vararg{Int}}, zero) where {T}
   return SparseArrayDOK{T}(dims, zero)
+end
+
+# Base `AbstractArray` interface
+Base.size(a::SparseArrayDOK) = a.dims
+
+function Base.similar(a::SparseArrayDOK, elt::Type, dims::Tuple{Vararg{Int}})
+  return SparseArrayDOK{elt}(undef, dims)
+end
+
+# SparseArrayInterface interface
+SparseArrayInterface.storage(a::SparseArrayDOK) = a.data
+
+getindex_zero_function(a::SparseArrayDOK) = a.zero
+
+# Conversion
+# TODO: Make these more generic to `AbstractSparseArray`.
+function SparseArrayDOK{T,N,Zero}(a::SparseArrayDOK{T,N,Zero}) where {T,N,Zero}
+  return copy(a)
+end
+
+function Base.convert(
+  ::Type{SparseArrayDOK{T,N,Zero}}, a::SparseArrayDOK{T,N,Zero}
+) where {T,N,Zero}
+  return a
+end
+
+SparseArrayDOK(a::AbstractArray) = SparseArrayDOK{eltype(a)}(a)
+
+SparseArrayDOK{T}(a::AbstractArray) where {T} = SparseArrayDOK{T,ndims(a)}(a)
+
+function SparseArrayDOK{T,N}(a::AbstractArray) where {T,N}
+  return SparseArrayInterface.sparse_convert(SparseArrayDOK{T,N}, a)
 end
