+++ NDTensors/src/lib/SparseArrayDOKs/test/Project.toml
@@ -1,0 +1,2 @@
+[deps]
+NDTensors = "23ae76d9-e61a-49c4-8f12-3f1a16adf9cf"
